Stats and UI

+++ Assets/1_Content/Scripts/Runtime/Components/Damage/IDamageProcessor.cs
@@ -4,6 +4,6 @@ namespace NJN.Runtime.Components
 {
     public interface IDamageProcessor : IComponent
     {
-        public void DealDamage(IDamagable damagable, int damageAmount);
+        public void DealDamage(IDamagable damagable, float damageAmount);
     }
 }
